@@ -14,15 +14,16 @@ classifiers = [
   "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
   "Operating System :: MacOS :: MacOS X",
   "Operating System :: POSIX :: Linux",
-  "Programming Language :: Python :: 3.8",
   "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
   "Programming Language :: Python :: 3.11",
   "Topic :: Utilities"
 ]
 dependencies = [
   "babel",
-  "backports.zoneinfo; python_version < '3.9'",
   "tabulate",
   "termcolor",
   "tomli; python_version < '3.11'",
@@ -42,10 +43,6 @@ test = ["pytest", "pytest-cov"]
 [project.scripts]
 multizone = "multizone:main"
 
-[tool.black]
-line-length = 108
-target-version = ['py38']
-
 [tool.coverage.run]
 include = ["multizone.py"]
 
@@ -53,18 +50,32 @@ include = ["multizone.py"]
 max-line-length = "108"
 
 [tool.pylint.message_control]
-disable = "R0912"
+disable = "R0912,R0914"
 
 [tool.ruff]
+line-length = 108
+target-version = "py39"
+
+[tool.ruff.lint]
 fixable = ["D", "I", "COM"]
 ignore = [
   "S101",  # assert
   "D213",  # multi-line-summary-second-line incompatible with multi-line-summary-first-line
   "D203"  # one-blank-line-before-class" incompatible with no-blank-line-before-class
 ]
-line-length = 108
 select = ["ALL"]
-target-version = "py38"
 
 [tool.ruff.per-file-ignores]
 "test_mz.py" = ["PLR2004"]
+
+[tool.tox]
+legacy_tox_ini = """
+[tox]
+envlist = py3{9,10,11,12,13}
+isolated_build = True
+
+[testenv]
+usedevelop=True
+deps = .[test]
+commands = pytest
+"""
